chore: revert font test and create new model and relation test file
